+++ pyccel/ast/bind_c.py
@@ -6,18 +6,49 @@
 
 from pyccel.ast.core import FunctionCall, Module
 from pyccel.ast.core import FunctionAddress
-from pyccel.ast.core import FunctionDef, BindCFunctionDef
+from pyccel.ast.core import FunctionDef
 from pyccel.ast.core import Assign
 from pyccel.ast.core import Import
 from pyccel.ast.core import AsName
 from pyccel.ast.variable import Variable
 
 __all__ = (
+    'BindCFunctionDef',
     'as_static_module',
     'as_static_function',
     'as_static_function_call',
     'sanitize_arguments',
 )
+
+class BindCFunctionDef(FunctionDef):
+    """
+    Contains the c-compatible version of the function which is
+    used for the wrapper.
+    As compared to a normal FunctionDef, this version contains
+    arguments for the shape of arrays. It should be generated by
+    calling ast.bind_c.as_static_function
+
+    Parameters
+    ----------
+    *args : See FunctionDef
+
+    original_function : FunctionDef
+        The function from which the c-compatible version was created
+    """
+    __slots__ = ('_original_function',)
+    _attribute_nodes = (*FunctionDef._attribute_nodes, '_original_function')
+
+    def __init__(self, *args, original_function, **kwargs):
+        self._original_function = original_function
+        super().__init__(*args, **kwargs)
+
+    @property
+    def name(self):
+        return str(self._name).lower()
+
+    @property
+    def original_function(self):
+        return self._original_function
 
 #=======================================================================================
 def sanitize_arguments(args):
@@ -189,3 +220,6 @@ def as_static_function_call(func, mod_name, name=None, imports = None):
     static_func = as_static_function(new_func, name)
 
     return static_func
+
+#=======================================================================================
+
