+++ pyccel/ast/core.py
@@ -57,7 +57,6 @@ __all__ = (
     'Duplicate',
     'DoConcurrent',
     'EmptyNode',
-    'BindCFunctionDef',
     'For',
     'ForIterator',
     'FunctionCall',
@@ -2995,36 +2994,6 @@ class SympyFunction(FunctionDef):
     """Represents a function definition."""
     __slots__ = ()
 
-
-class BindCFunctionDef(FunctionDef):
-    """
-    Contains the c-compatible version of the function which is
-    used for the wrapper.
-    As compared to a normal FunctionDef, this version contains
-    arguments for the shape of arrays. It should be generated by
-    calling ast.bind_c.as_static_function
-
-    Parameters
-    ----------
-    *args : See FunctionDef
-
-    original_function : FunctionDef
-        The function from which the c-compatible version was created
-    """
-    __slots__ = ('_original_function',)
-    _attribute_nodes = (*FunctionDef._attribute_nodes, '_original_function')
-
-    def __init__(self, *args, original_function, **kwargs):
-        self._original_function = original_function
-        super().__init__(*args, **kwargs)
-
-    @property
-    def name(self):
-        return str(self._name).lower()
-
-    @property
-    def original_function(self):
-        return self._original_function
 
 
 class ClassDef(ScopedNode):
